network: Split FD select code to platform-specific C files
Roll out two do_select() functions, one for Windows and one for UNIX, to
make the code simpler.
